requirements.txt
certifi==2021.10.8
wincertstore==0.2
mlflow
pandas
numpy
PyYAML
Flask
Flask-Cors
gunicorn
Jinja2
matplotlib
dnspython==1.16.0

### to install local packages -
-e . 